feat(web): use Umami for GitHub Pages demo instead of GA4
Load cloud.umami.is on the fragment-mode Pages build only; track hash
navigation for flow deep links. Local dev and CLI deploys unchanged.
