+++ .github/workflows/pages.yml
@@ -38,9 +38,8 @@ jobs:
           # https://naorsabag.github.io/openhop/ — is the asset base path.
           VITE_BASE: /openhop/
           VITE_FRAGMENT_MODE: "1"
-          # GA4 for the public Pages demo only. Add repo secret GA_MEASUREMENT_ID
-          # (Settings → Secrets → Actions) with your G-XXXXXXXXXX ID. Unset = no analytics.
-          VITE_GA_MEASUREMENT_ID: ${{ secrets.GA_MEASUREMENT_ID }}
+          # Umami for the public Pages demo only (cloud.umami.is). Unset = no analytics.
+          VITE_UMAMI_WEBSITE_ID: "139df029-9f0e-4323-a08e-46a6fe7674a6"
       - uses: actions/configure-pages@v6
         with:
           enablement: true
